+++ packages/client/src/App.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react'
 import { Provider } from 'react-redux';
 import { type AnyAction, configureStore, type ThunkMiddleware } from '@reduxjs/toolkit';
 import type { ToolkitStore } from '@reduxjs/toolkit/dist/configureStore'
-import { userSlice } from './slices/userSlice'
+import { userSlice } from './slices/user-slice'
 
 const appContent = 'Пользователь не найден!'
