+++ packages/client/src/enums/api.ts
@@ -1,0 +1,6 @@
+export enum ERequestMethods {
+  GET = 'GET',
+  POST = 'POST',
+  PUT = 'PUT',
+  DELETE = 'DELETE',
+}
