+++ packages/client/src/store.ts
@@ -9,7 +9,7 @@ import { configureStore } from '@reduxjs/toolkit'
 
 import friendsReducer from './slices/friendsSlice'
 import ssrReducer from './slices/ssrSlice'
-import userReducer from './slices/userSlice'
+import userReducer from './slices/user-slice'
 
 // Глобально декларируем в window наш ключик
 // и задаем ему тип такой же как у стейта в сторе
