+++ packages/client/src/types/common.ts
@@ -1,0 +1,1 @@
+export type TNullable<T> = T | null;
