+++ packages/client/src/types/index.ts
@@ -2,4 +2,4 @@ export type {
   StringSchemaValueType,
 } from './yup-schema';
 export type { IUser, IUserPassword } from './user';
-
+export type { TNullable } from './common';
